2020'den sonraki studentID'lerin int range'inden dışarı çıkması sorunu çözüldü

+++ create.sql
@@ -52,7 +52,7 @@ CREATE TABLE users (
 );
 
 CREATE TABLE student (
-    studentID INT NOT NULL,
+    studentID BIGINT NOT NULL,
     schoolEnrollmentDate DATE NOT NULL,
     semesterECTS INT,
     PRIMARY KEY (ID),
